requirements.txt
aiogram>=3.4.1
SQLAlchemy>=2.0.0
asyncpg
aiohttp
python-dotenv

